requirements.txt
streamlit==1.44.1
cohere==5.15.0
python-dotenv==1.1.0
